divided the calculations of various coordinates systems in coordinates and added the geo coordination call there removing it from radar_pos.py
